pyproject.toml
[build-system]
requires = ["setuptools>=68", "build", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "green-white-agent"
version = "0.1.0"
description = "Green agent for Terminal-Bench integration with A2A protocol support"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Green Agent Team", email = "team@greenagent.com" }]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "fastapi==0.119.1",
    "httpx>=0.27.0",
    "litellm>=1.0.0",
    "openai==2.6.0",
    "pydantic>=2.12.3",
    "pydantic-settings==2.11.0",
    "pydantic_core==2.41.4",
    "python-dotenv==1.1.1",
    "PyYAML==6.0.3",
    "requests==2.32.5",
    "typer>=0.19.2",
    "uvicorn==0.38.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
full = ["terminal-bench"]

[tool.setuptools.packages.find]
where = ["."]
